Remove broadcast option
Setting this on a network that uses DHCP is a Bad Idea™, because it will
wake the computer up almost immediately after shutting down, presumably
because someone broadcasted a DHCP query. It could be something else,
though, I don't know, as I am at the whims of unnamed firmware authors that
have actually implemented the logic behind it.