Fix the Emacs situation
For more than 10 years, I happily used the terse oneliner that has now
been removed, because it stopped working for unknown reasons. A better
solution is to always start the Emacs daemon on system startup, and
use yet another terse command to find out whether the user has opened
any frames (windows) yet. If not, `e` will do that, else `e` will open
the file in the existing frame. Which is exactly the same behavior as
the old oneliner, except that it works again. Old habits die hard!